components folder reorganized
